Simplify mapbox_layer_upload, guarantee serialization order, and reduce db calls.
This change reduces the number of database calls to retrieve the mapbox data
from a maximum of 4 to exactly 1.

This change also adds a constant to MapboxLayerForm that is used for determining
the order in which fields are serialized or deserialized. This prevents needing a
migration for changing the field ordering and allows currently serialized data
to be appropriately deserialized by only changing the serialization order
constant. A similar change could be made which uses the declared_fields
attribute of MapboxLayerForm rather than defining a new constant, but this would
still leave the serialization process dependent upon the field declaration
order.